Opted to use full namespace instead.
The reason is that it will make it easier to move files around for maintenance.  Using the full namespaces allows for references to be found and updated quicker using find and replace on the whole project rather than having to check each file for possible matches.  It should reduce errors since PHP is not a compiled language (at least in most implementations).